requirements.txt
PyYAML>=6.0.2
PyMuPDF>=1.24.0
Pillow>=10.0.0
